fix merges FWO.Basics
