--- roles/GlobalFuncs/FWO.GlobalFunctions.csproj
@@ -1,13 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-	<PropertyGroup>
-		<TargetFramework>net8.0</TargetFramework>
-		<ImplicitUsings>enable</ImplicitUsings>
-		<Nullable>enable</Nullable>
-	</PropertyGroup>
-
-	<ItemGroup>
-		<PackageReference Include="IPAddressRange" Version="6.0.0" />
-	</ItemGroup>
-
-</Project>
