+++ roles/lib/files/FWO.Basics/FWO.Basics.csproj
@@ -6,4 +6,8 @@
     <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
 
+  <ItemGroup>
+    <PackageReference Include="IPAddressRange" Version="6.0.0" />
+  </ItemGroup>
+
 </Project>
