+++ roles/lib/files/FWO.Report/ReportAppRules.cs
@@ -101,8 +101,8 @@ namespace FWO.Report
         {
             List<ModellingAppServer> appServers = await apiConnection.SendQueryAsync<List<ModellingAppServer>>(ModellingQueries.getAppServers, 
                 new { appId = Query.SelectedOwner?.Id });
-            ownerIps = [.. appServers.ConvertAll(s => new IPAddressRange(IPAddress.Parse(DisplayBase.StripOffNetmask(s.Ip)),
-                IPAddress.Parse(DisplayBase.StripOffNetmask(s.IpEnd != "" ? s.IpEnd : s.Ip))))];
+            ownerIps = [.. appServers.ConvertAll(s => new IPAddressRange(IPAddress.Parse(s.Ip.StripOffNetmask()),
+                IPAddress.Parse((s.IpEnd != "" ? s.IpEnd : s.Ip).StripOffNetmask())))];
         }
 
         private (List<NetworkLocation>, List<NetworkLocation>) CheckNetworkObjects(NetworkLocation[] objList)
@@ -156,8 +156,8 @@ namespace FWO.Report
             foreach(var ownerIpRange in ownerIps)
             {
                 if(obj.IP != null &&
-                    ComplianceNetworkZone.OverlapExists(new IPAddressRange(IPAddress.Parse(DisplayBase.StripOffNetmask(obj.IP)),
-                    IPAddress.Parse(DisplayBase.StripOffNetmask(obj.IpEnd != null && obj.IpEnd != "" ? obj.IpEnd : obj.IP))), ownerIpRange))
+                    ComplianceNetworkZone.OverlapExists(new IPAddressRange(IPAddress.Parse(obj.IP.StripOffNetmask()),
+                    IPAddress.Parse((obj.IpEnd != null && obj.IpEnd != "" ? obj.IpEnd : obj.IP).StripOffNetmask())), ownerIpRange))
                 {
                     return true;
                 }
