+++ roles/middleware/files/FWO.Middleware.Server/AreaIpDataImport.cs
@@ -7,8 +7,6 @@ using FWO.Config.Api;
 using System.Text.Json;
 using NetTools;
 using System.Reactive.Subjects;
-using System.Linq;
-using FWO.GlobalFunctions;
 
 namespace FWO.Middleware.Server
 {
