+++ roles/ui/files/FWO.UI/Pages/Monitoring/MonitorImportStatus.razor
@@ -3,7 +3,6 @@
 @using FWO.Api.Data
 @using FWO.Api.Client
 @using FWO.Config.Api
-@using FWO.GlobalFunctions
 @using FWO.Ui.Shared
 
 @page "/monitoring/import_status"
