+++ roles/ui/files/FWO.UI/Pages/NetworkModelling/ManualAppServer.razor
@@ -1,5 +1,5 @@
 ﻿@using FWO.Config.Api
-@using FWO.GlobalFunctions
+@using FWO.Basics
 @using FWO.Ui.Display
 @using System.Text.Json
 @using System.Net
