+++ roles/ui/files/FWO.UI/Pages/Reporting/Schedule.razor
@@ -1,5 +1,5 @@
 ﻿@using FWO.Config.Api
-@using FWO.GlobalFunctions
+@using FWO.Basics
 @using FWO.Report.Filter
 
 @implements IDisposable
