+++ roles/ui/files/FWO.UI/Pages/Request/DisplayImplTaskTable.razor
@@ -2,7 +2,6 @@
 @using FWO.Basics
 @using FWO.Api.Data
 @using FWO.Config.Api
-@using FWO.GlobalFunctions
 @using FWO.Ui.Services
 
 @attribute [Authorize(Roles = $"{Roles.Admin}, {Roles.Requester}, {Roles.Approver}, {Roles.Planner}, {Roles.Implementer}, {Roles.Reviewer}, {Roles.Auditor}, {Roles.FwAdmin}")]
