+++ roles/ui/files/FWO.UI/Pages/Settings/SettingsCredentials.razor
@@ -5,7 +5,6 @@
 @using FWO.Api.Client.Queries
 @using FWO.Encryption
 @using FWO.Config.Api
-@using FWO.GlobalFunctions
 @using FWO.Middleware.Client
 
 @page "/settings/credentials"
