+++ roles/ui/files/FWO.UI/Pages/Settings/SettingsGateways.razor
@@ -3,7 +3,6 @@
 @using FWO.Api.Client
 @using FWO.Api.Client.Queries
 @using FWO.Config.Api
-@using FWO.GlobalFunctions
 @using FWO.Middleware.Client
 
 @page "/settings/gateways"
