+++ roles/ui/files/FWO.UI/Pages/Settings/SettingsGroups.razor
@@ -3,7 +3,6 @@
 @using FWO.Basics
 @using FWO.Api.Data
 @using FWO.Api.Client
-@using FWO.GlobalFunctions
 @using FWO.Ui.Services
 @using FWO.Middleware.Client
 @using FWO.Config.Api
