+++ roles/ui/files/FWO.UI/Pages/Settings/SettingsLdap.razor
@@ -1,6 +1,5 @@
 ﻿@using BlazorTable
 @using System.Net
-@using FWO.GlobalFunctions
 @using RestSharp
 
 @using FWO.Basics
