+++ roles/ui/files/FWO.UI/Pages/Settings/SettingsManagements.razor
@@ -5,7 +5,6 @@
 @using FWO.Api.Client.Queries
 @using FWO.Config.Api
 @using FWO.DeviceAutoDiscovery
-@using FWO.GlobalFunctions
 @using FWO.Middleware.Client
 
 @page "/settings/managements"
