+++ roles/ui/files/FWO.UI/Pages/Settings/SettingsUsers.razor
@@ -4,7 +4,6 @@
 @using FWO.Api.Data
 @using FWO.Api.Client
 @using FWO.Api.Client.Queries
-@using FWO.GlobalFunctions
 @using FWO.Middleware.Client
 @using FWO.Config.Api
 @using FWO.Middleware.RequestParameters
