+++ roles/ui/files/FWO.UI/Services/ModellingAppServerHandler.cs
@@ -4,8 +4,6 @@ using FWO.Basics;
 using FWO.Api.Data;
 using FWO.Api.Client;
 using FWO.Api.Client.Queries;
-using FWO.GlobalFunctions;
-
 
 namespace FWO.Ui.Services
 {
