+++ roles/ui/files/FWO.UI/Shared/ConnectionTable.razor
@@ -1,5 +1,4 @@
 @using FWO.Config.Api
-@using FWO.GlobalFunctions
 @using FWO.Ui.Display
 @using FWO.Basics
 @using FWO.Api.Data
