+++ roles/ui/files/FWO.UI/Shared/ImportDetails.razor
@@ -2,7 +2,6 @@
 @using FWO.Basics
 @using FWO.Api.Data
 @using FWO.Api.Client
-@using FWO.GlobalFunctions
 
 @inject ApiConnection apiConnection
 @inject UserConfig userConfig
